Client: separated point definitions into a separate file
